pyproject.toml
[project]
name = "medhaai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "lxml>=6.0.0",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
